// generated/src/app/api-models/commission-determ-header.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CommissionDetermHeader {
  matrixDeterminant: string;
  matrixDescription: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}